# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Internship API",
    "description": "A  cool app to track your internship applications",
    "version": "1.0.0"
  },
  "host": "",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/dashboard": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/auth/google": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/auth/google/callback": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/auth/logout": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/internships/add": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/internships/": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      },
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/internships/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/internships/edit/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    }
  }
}